+++ Calculator/src/main/java/application/view/AsciiView.java
@@ -1,43 +1,43 @@
 package application.view;
 
-import application.calculator.OpType;
 import java.util.Scanner;
 import java.util.function.Consumer;
-import java.util.function.Function;
-
 
 /**
- * Skeleton code for a terminal based calculator that reads an expression from the user and
- * evaluates it and prints out the answer.
+ * View for a terminal based calculator that reads an expression from the user and evaluates it and
+ * prints out the answer. Expressions must be entered like ? 3 * ( 5 + 4 )
  * 
  * @author zlac318
  */
 public class AsciiView implements ViewInterface {
-  // The current question that the calculator must solve: entered like ?3*(5+4)
+
+  // The current question that the calculator must solve: entered like ? 3 * ( 5 + 4 )
   private String question;
-  private String answer;
-  private Function<String, String> evaluator;
 
   // This method will be injected so we can ask the controller to calculate
-  Runnable doCalculation = null;
+  Runnable doCalculation;
 
   // This method changes how the calculator will evaluate the question
-  Consumer<OpType> setCalculatorType = null;
-
+  Consumer<String> setCalculatorType;
 
   private void menu() {
     Scanner s = new Scanner(System.in);
     boolean finished = false;
     help();
-
     while (!finished && s.hasNext()) {
       String t = s.next();
       switch (t.toUpperCase().charAt(0)) {
         case 'C': // Ask the controller to calculate
-          System.out.println(evaluator.apply(question).toString());
+          doCalculation.run();
+          break;
+        case 'S':
+          setCalculatorType.accept("infix");
+          break;
+        case 'R':
+          setCalculatorType.accept("postfix");
           break;
         case '?': // Set current question
-          question = s.nextLine().replace("?", "");
+          question = s.nextLine();
           break;
         case 'Q':
           System.out.println("Bye");
@@ -52,7 +52,7 @@ public class AsciiView implements ViewInterface {
 
   private void help() {
     System.out.println("Use one of the following:");
-    System.out.println("  ?Expression - to set expression");
+    System.out.println("  ? Expression - to set expression");
     System.out.println("  C - to calculate");
     System.out.println("  S - change to a standard calculator");
     System.out.println("  R - change to a reverse polish calculator");
@@ -61,25 +61,55 @@ public class AsciiView implements ViewInterface {
 
   @Override
   public String getExpression() {
-    return "0";
+    return question;
   }
 
   @Override
   public void setAnswer(String answer) {
-    System.out.println("Answer is just around the corner");
+    System.out.println(answer);
   }
 
   @Override
-  public void addCalculateObserver(Function<String, String> evaluate) {
-    evaluator = evaluate;
+  public void addTypeObserver(Consumer<String> c) {
+    setCalculatorType = c;
   }
-
-  @Override
-  public void addTypeObserver(Consumer<OpType> c) {}
 
   @Override
   public void startView() {
     menu();
   }
+
+  @Override
+  public void addCalculateObserver(Runnable function) {
+    doCalculation = function;
+  }
+
+  /////////////////////////////////////////////////////////////////////////////////
+  // Block for creating an instance variable for others to use - singleton.
+
+  /**
+   * A private constructor for the singleton View. Constructs the view.
+   */
+  private AsciiView() {
+    doCalculation = null;
+    setCalculatorType = null;
+  }
+
+  private static volatile AsciiView instance = new AsciiView();
+
+  /**
+   * This is a Singleton View made available through this method.
+   * 
+   * @return the single object representing this view
+   */
+  public static synchronized AsciiView getInstance() {
+    if (instance == null) {
+      instance = new AsciiView();
+    }
+    return instance;
+  }
+  // End of special block
+  /////////////////////////////////////////////////////////////////////////////////
+
 
 }